+++ MatrixSDK/MXSession.m
@@ -879,11 +879,8 @@ typedef void (^MXOnResumeDone)(void);
         }];
     }
 
-    // Get wellknown data only at the login time
-    if (!self.homeserverWellknown)
-    {
-        [self refreshHomeserverWellknown:nil failure:nil];
-    }
+    // Refresh wellknown data
+    [self refreshHomeserverWellknown:nil failure:nil];
 }
 
 - (NSString *)syncFilterId